added the second class user
